requirements.txt
pandas 
openpyxl
pyxlsb 
soundfile 
torchaudio 
pydub 
transformers
openai-whisper
pyannote.audio
jiwer
flask
dotenv